immg-master/peoplemanagement/requirements.txt
dataclasses==0.6
Django==2.2.5
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
djongo==1.2.36
Pillow==6.1.0pyt
pytz==2019.2
sqlparse==0.2.4
